Reorganize codebase, add schemas for - WebhookNotification - ListWebhooksResponse - WebhookSpecification
