+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "airtable-webhook-types",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "description": "Typescript and Zod types for Airtable webhooks",
   "license": "MIT",
   "author": "Dominic Roy-Stang",
